agent/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.2
google-generativeai>=0.3.0
langgraph>=0.0.20
transformers>=4.35.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
torch>=2.0.0
numpy>=1.24.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0 